requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pypdf
reportlab
segno
arabic-reshaper
python-bidi
python-multipart
python-dotenv
supabase
gunicorn
bcrypt
python-jose[cryptography]
websockets>=13.0
email-validator
pytest
